快速跑轨

+++ master/HAL-Version/Application/TaskLayer/monitor_task.c
@@ -22,19 +22,19 @@
 /* Private functions ---------------------------------------------------------*/
 static void device_heart_beat(void)
 {
-	rc_sensor.heart_beat(&rc_sensor);
-	imu_sensor.heart_beat(&imu_sensor);
-	motor[CHASSIS].heart_beat(&motor[CHASSIS]);
-	motor[DIAL].heart_beat(&motor[DIAL]);
-	motor[GIMBAL_PITCH].heart_beat(&motor[GIMBAL_PITCH]);
-	motor[GIMBAL_YAW].heart_beat(&motor[GIMBAL_YAW]);
-	vision_sensor.heart_beat(&vision_sensor);
+    rc_sensor.heart_beat(&rc_sensor);
+    imu_sensor.heart_beat(&imu_sensor);
+    motor[CHASSIS].heart_beat(&motor[CHASSIS]);
+    motor[DIAL].heart_beat(&motor[DIAL]);
+    motor[GIMBAL_PITCH].heart_beat(&motor[GIMBAL_PITCH]);
+    motor[GIMBAL_YAW].heart_beat(&motor[GIMBAL_YAW]);
+    vision_sensor.heart_beat(&vision_sensor);
 }
 
 static void system_led_flash(void)
 {
-    static int cnt=0;
-    static int colour=1;
+    static int16_t cnt=0;
+    static int16_t colour=1;
     if(colour==1)
     {
         LED_ORANGE_ON();
@@ -55,10 +55,10 @@ static void system_led_flash(void)
 
 static void device_get(void)
 {
-	imu_sensor.update(&imu_sensor);
-	path_sensor.update(&path_sensor);
-	path_sensor.check(&path_sensor);
-	Update_Gimbal_Angle_Queue(motor[GIMBAL_YAW].info->angle_sum);//更新云台队列
+    imu_sensor.update(&imu_sensor);
+    path_sensor.update(&path_sensor);
+    path_sensor.check(&path_sensor);
+    Update_Gimbal_Angle_Queue(motor[GIMBAL_YAW].info->angle_sum);//更新云台队列
 }
 /* Exported functions --------------------------------------------------------*/
 /**
@@ -66,11 +66,11 @@ static void device_get(void)
  */
 void StartMonitorTask(void const * argument)
 {
-	for(;;)
-	{
-		system_led_flash();
-		device_heart_beat();
-		device_get();
-		osDelay(1);
-	}
+    for(;;)
+    {
+        system_led_flash();
+        device_heart_beat();
+        device_get();
+        osDelay(1);
+    }
 }
